Rename "All sales" -> "Investable sales"
