+++ src/modules-arch/schemes/reit-en.json
@@ -111,7 +111,7 @@
       "overview-title": "Opportunity overview",
       "asset-code-label": "Opportunity code",
       "new-sale": "New opportunity",
-      "all-sales": "All opportunities",
+      "investable-sales": "Investable opportunities",
       "my-sales": "My opportunities"
     },
     "trade-open-offers": {
