+++ src/modules-arch/schemes/vanilla.js
@@ -37,7 +37,7 @@ import { PreIssuanceFormModule } from '@/vue/modules/pre-issuance-form/module'
 import { TransferDrawerPseudoModule } from '@/modules-arch/pseudo-modules/transfer-drawer-pseudo-module'
 import { CreateSaleFormModule } from '@modules/create-sale-form/module'
 import { DashboardChartPseudoModule } from '@/modules-arch/pseudo-modules/dashboard-chart-pseudo-module'
-import { SalesListPageModule } from '@/vue/pages/sales/all-sales-page-module'
+import { SalesListPageModule } from '@/vue/pages/sales/investable-sales-page-module'
 import { SalesListOwnedPageModule } from '@/vue/pages/sales/user-owned-sales-page-module'
 import { SaleCampaignViewerPageModule } from '@/vue/pages/sale-details/sale-campaign-viewer-page-module'
 import { SaleStateWidgetModule } from '@/vue/pages/sale-details/sale-sate-widget-module'
@@ -209,7 +209,7 @@ export default {
           new SalesListPageModule({
             routerEntry: {
               path: '/sales/all',
-              name: vueRoutes.allSales.name,
+              name: vueRoutes.investableSales.name,
               props: {
                 default: true,
                 isUserSales: false,
