+++ src/vue-router/routes.js
@@ -11,7 +11,7 @@ export const vueRoutes = Object.freeze({
   sales: { name: 'app.sales' },
   saleDetails: { name: 'app.sale-details' },
   saleCampaign: { name: 'app.sale-campaign' },
-  allSales: { name: 'app.all-sales' },
+  investableSales: { name: 'app.investable-sales' },
   userOwnedSales: { name: 'app.sales-user' },
   signup: { name: 'signup' },
   recovery: { name: 'recovery' },
