+++ src/vue/pages/sales-page-module.js
@@ -1,6 +1,6 @@
 import { PageModuleDescriptor } from '@/modules-arch/page-module-descriptor'
 import { CreateSaleFormModule } from '@modules/create-sale-form/module'
-import { SalesListPageModule } from '@/vue/pages/sales/all-sales-page-module'
+import { SalesListPageModule } from '@/vue/pages/sales/investable-sales-page-module'
 import { SalesListOwnedPageModule } from '@/vue/pages/sales/user-owned-sales-page-module'
 import { CreateOpportunityModule } from '../modules/create-opportunity/module'
