+++ src/vue/pages/sales/user-owned-sales-page-module.js
@@ -4,7 +4,7 @@ export class SalesListOwnedPageModule extends PageModuleDescriptor {
   constructor (opts = {}) {
     super({
       ...opts,
-      importComponentFn: _ => import('@/vue/pages/sales/AllSales'),
+      importComponentFn: _ => import('@/vue/pages/sales/SalesList'),
     })
   }
 }
